Factor out CCoinsView based AreInputsStandard/IsWitnessStandard This will allow IsStandard checks without depending on the mempool, UTXO set
