+++ src/test/transaction_tests.cpp
@@ -15,6 +15,7 @@
 #include "keystore.h"
 #include "validation.h"
 #include "policy/policy.h"
+#include "policy/inputs.h"
 #include "script/script.h"
 #include "script/sign.h"
 #include "script/script_error.h"
